requirements.txt
openai>=1.0.0
llama-index>=0.10.0
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma
chromadb
python-docx
streamlit
